refactor: move bulk scheme to seperate file
