+++ libs/kid_bulk_microphysics/__init__.py
@@ -1,0 +1,18 @@
+"""
+Copyright (c) 2024 MPI-M, Clara Bayley
+
+----- Microphysics Test Cases -----
+File: __init__.py
+Project: kid_bulk_scheme
+Created Date: Monday 2nd September 2024
+Author: Clara Bayley (CB)
+Additional Contributors:
+-----
+Last Modified: Monday 2nd September 2024
+Modified By: CB
+-----
+License: BSD 3-Clause "New" or "Revised" License
+https://opensource.org/licenses/BSD-3-Clause
+-----
+File Description:
+"""
